+++ examples/interval.py
@@ -6,7 +6,7 @@
 # Purpose:  General `Interval` demonstration (versus `asyncio.sleep`).
 #
 # Created:  25th July 2025
-# Updated:  27th June 2026
+# Updated:  12th July 2026
 #
 # Author:   Matthew Wilson
 #
@@ -16,18 +16,19 @@
 # ######################################################################## #
 
 
+import asyncio
+import sys
+import time
+
+import diagnosticism as d
+import diagnosticism.severity as sev
+
 from asynkio.time import (
     Duration,
     Instant,
     Interval,
     MissedTickBehaviour,
 )
-
-import asyncio
-import diagnosticism as d
-import diagnosticism.severity as sev
-import sys
-import time
 
 
 def _delta_as_ns(
@@ -39,10 +40,11 @@ def _delta_as_ns(
 
     return delta
 
+
 def _clipped_delta_as_ns(
     f: Instant,
     t: Instant,
-    quantum_ns : int
+    quantum_ns: int,
 ) -> str:
 
     delta = int(t) - int(f)
@@ -52,9 +54,9 @@ def _clipped_delta_as_ns(
 
 
 async def run_asyncio_sleep(
-    label : str,
-    timeout_ms : int,
-    synchronous_sleep_ms : int,
+    label: str,
+    timeout_ms: int,
+    synchronous_sleep_ms: int,
 ):
 
     d.trace()
@@ -73,15 +75,18 @@ async def run_asyncio_sleep(
 
         t2 = Instant.now()
 
-        d.log(sev.NOTICE, f"{label} slept for {_delta_as_ns(t1, t2):,}; ∆={_clipped_delta_as_ns(t0, t2, timeout_ms * 1_000_000)}")
+        d.log(
+            sev.NOTICE,
+            f"{label} slept for {_delta_as_ns(t1, t2):,}; ∆={_clipped_delta_as_ns(t0, t2, timeout_ms * 1_000_000)}",
+        )
 
         t1 = t2
 
 
 async def run_asynkio_interval(
-    label : str,
-    timeout_ms : int,
-    synchronous_sleep_ms : int,
+    label: str,
+    timeout_ms: int,
+    synchronous_sleep_ms: int,
 ):
 
     d.trace()
@@ -91,8 +96,6 @@ async def run_asynkio_interval(
     t0 = Instant.now()
 
     t1 = t0
-
-    count = 0
 
     while True:
 
@@ -104,13 +107,16 @@ async def run_asynkio_interval(
 
         t2 = Instant.now()
 
-        d.log(sev.NOTICE, f"{label} slept for {_delta_as_ns(t1, t2):,}; ∆={_clipped_delta_as_ns(t0, t2, timeout_ms * 1_000_000)}")
+        d.log(
+            sev.NOTICE,
+            f"{label} slept for {_delta_as_ns(t1, t2):,}; ∆={_clipped_delta_as_ns(t0, t2, timeout_ms * 1_000_000)}",
+        )
 
         t1 = t2
 
 
 async def heartbeat(
-    missed_tick_behaviour : MissedTickBehaviour,
+    missed_tick_behaviour: MissedTickBehaviour,
 ):
 
     d.trace()
@@ -136,28 +142,31 @@ async def heartbeat(
 
         t2 = Instant.now()
 
-        d.log(sev.INFO, f"❤️  #{count} : {t2 - t0}, {t2 - t1}, {Duration.from_nanos((t2 - t0).as_nanos() % 1_000_000_000)}")
+        d.log(
+            sev.INFO,
+            f"❤️  #{count} : {t2 - t0}, {t2 - t1}, {Duration.from_nanos((t2 - t0).as_nanos() % 1_000_000_000)}",
+        )
 
         t1 = t2
 
 
 async def main(
-    missed_tick_behaviour : MissedTickBehaviour,
+    missed_tick_behaviour: MissedTickBehaviour,
 ):
 
     d.trace()
 
     await asyncio.gather(
         heartbeat(missed_tick_behaviour),
-        run_asyncio_sleep(   "asyncio.sleep      ", 5_000, synchronous_sleep_ms=0),
+        run_asyncio_sleep("asyncio.sleep      ", 5_000, synchronous_sleep_ms=0),
         run_asynkio_interval("asynkio.interval a1", 2_500, synchronous_sleep_ms=0),
     )
 
-def show_help(
-    file
-):
+
+def show_help(file):
 
     d.abort(f"USAGE: {d.get_program_name()} {{ burst | delay | skip }}", file=file, show_program_name=False)
+
 
 if __name__ == "__main__":
 
@@ -188,3 +197,4 @@ if __name__ == "__main__":
             except KeyboardInterrupt:
 
                 pass
+